Update README.md with instructions to install to the User Library
It's not recommended to change files in Ableton Live's application bundle. Especially on macOS this can mess with code signing and notarization. Live can read external MIDI Remote Scripts from the 'Remote Scripts' folder in the User Library. This directory does not exist by default, so the user might need to create it themselves.